revert indexes to uint32_t

+++ lib/DB_File.h
@@ -10,6 +10,6 @@ typedef struct{
 }t_datablock;
 
 int Create_DB(dbc *db, char filename[]);
-long create_index_file(dbc* db, meta_t* meta, uint64_t nb, int (*setSlot)(void*, void*), t_datablock* i_block, t_datablock* t_block);
+long create_index_file(dbc* db, meta_t* meta, uint32_t nb, int (*setSlot)(void*, void*), t_datablock* i_block, t_datablock* t_block);
 
 #endif // DB_FILE_H_INCLUDED
